+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 ------------------
 
 * BODY-53: Additional metadata from image (ericof).
+* LEICA-50: Add custom resource event types for models: customer, comments, project. (rudaporto)
 
 1.0.0
 -----